+++ Robust.Shared/Toolshed/ToolshedManager.Permissions.cs
@@ -39,10 +39,7 @@ public sealed partial class ToolshedManager
             if (_net.IsClient)
                 throw new NotImplementedException("Toolshed is not yet ready for client-side use.");
 #endif
-
-            if (!Started)
-                Startup();
-
+            _defaultEnvironment ??= new();
             return _defaultEnvironment;
         }
     }